requirements.txt
tensorflow
opencv-python
numpy
google-cloud-storage
h5py
